requirements.txt
Django==5.0.7
djangorestframework==3.15.2
psycopg2>=2.9.8
python-dotenv==1.0.0
djangorestframework-simplejwt==5.3.1

# flake8>=3.6.0,<3.7.0